feat: [IOBP-689,IOBP-699] Add new payment method management type handling (#5860)
## Short description
This PR adds management of REDIRECT-type payment methods. If a payment
method is of type REDIRECT, during the generation of the authorization
request it must be passed as `detailType` to `redirect`.

## List of changes proposed in this pull request
- Added `REDIRECT` as a valid payment method
- Add the handling of the payment method management type in the
`handleWalletPaymentAuthorization` saga. If the method management is
REDIRECT, it means that we have to provide
`RedirectDetailTypeEnum.redirect` as `detailType` attribute

## How to test
- Checkout the dev-server from this PR:
https://github.com/pagopa/io-dev-api-server/pull/390
- You should be able to see a new payment method (i.e. PostePay)
- Selecting that payment method, when you reach the authentication
request (when you're in the confirmation screen), you should check in
the request that `detailType` is `redirect`

## Preview


https://github.com/pagopa/io-app/assets/34343582/b4594d9b-5321-48d9-ad06-c1026e7162ce