[IOCOM-1557] New payment flow from a message's details (#5899)
## Short description
This PR adds the new payment flow when starting from a message (standard
or SEND). It is triggered only if the related feature flag is active.

The payment code that handles new payment flow does not support SEND
messages or new DS ones (it is compatible only with the legacy standard
message). In order to meet the deadline, such code had to be duplicated
and changed to support both SEND and new standard messages. It is also
uncovered by automatic testing.

## List of changes proposed in this pull request
-
`ts/features/payments/checkout/tempWorkaround/pagoPaPaymentWorkaround.ts`:
duplicated code to support SEND and new standard messages (the main
difference is that it does not rely upon being callend inside a React
Component, so not having a direct dependency upon useDispatch and
useNavigation hooks)
- `ts/features/messages/utils/index.ts`: entry point for the new flow
for SEND and new standard messages
- `ts/features/messages/components/MessageDetail/PaymentButton.tsx`:
entry point for the old standard messages
- New entry point callers (SEND messages and new standard ones):
-
`ts/features/messages/components/MessageDetail/MessageDetailsPaymentButton.tsx`
- `ts/features/messages/components/MessageDetail/MessagePaymentItem.tsx`
  - `ts/features/pn/components/MessageFooter.tsx`

## How to test
Using the io-dev-api-server on this branch
[IOBP-698-mock-redirect-method](https://github.com/pagopa/io-dev-api-server/tree/IOBP-698-mock-redirect-method),
generate messages with payment and SEND messages with payment(s). For
each combination of the new DS and the new Wallet, check that the flow
work as intended. Exhaustive list of use-cases is:
- DS off, Wallet off, start payment from the CTA in a standard message;
- DS off, Wallet off, start payment from the payment list in a SEND
message;
- DS off, Wallet off, start payment from the payment button in a SEND
message;
- DS on, Wallet off, start payment from the CTA in a standard message;
- DS on, Wallet off, start payment from the payment list in a standard
message;
- DS on, Wallet off, start payment from the payment list in a SEND
message;
- DS on, Wallet off, start payment from the payment button in a SEND
message;
- DS off, Wallet on, start payment from the CTA in a standard message;
- DS off, Wallet on, start payment from the payment list in a SEND
message;
- DS off, Wallet on, start payment from the payment button in a SEND
message;
- DS on, Wallet on, start payment from the CTA in a standard message;
- DS on, Wallet on, start payment from the payment list in a standard
message;
- DS on, Wallet on, start payment from the payment list in a SEND
message;
- DS on, Wallet on, start payment from the payment button in a SEND
message;