chore: [IOCOM-1525] PAID badge, new DS (#5877)
⚠️ This PR depends on #5873 ⚠️

## Short description
This PR enables the PAID badge on messages in the new DS messages' home.
![Simulator Screenshot - iPhone 15 - 2024-06-19 at 14 21
13](https://github.com/pagopa/io-app/assets/5150343/36ca20db-b47a-41b1-8297-8ae9d647f2d5)

## List of changes proposed in this pull request
- `MessageListItem`: support for `success` badgeVariant
- `WrappedMessageListItem`: business logic for `badgeVariant` and
`badgeText`
- `allPaginated` refactored `isNoticePaidSelector` into
`isPaymentMessagewithPaidNoticeSelector`

## How to test
Using the io-dev-api-server, generate some messages with and without
payment. Make sure to enable both DS and new Home. Select a message with
a payment and pay it. The related message's cell should display the PAID
badge.
Be aware that SEND messages with payments do not display the PAID badge.