chore(IT Wallet): [SIW-1382] Remove TINIT prefix from fiscal code claim visualisation (#6030)
## Short description
This PR removes the `TINIT` prefix from the fiscal code claim during the
visualisation.

## List of changes proposed in this pull request
- Adds a `FiscalCodeClaim` decoder which uses a regex to parse a fiscal
code with the `TINIT` prefix;
- Adds the aforementioned decoder in both `ItwRequiredClaimsList` and
`ItwCredentialClaim` which render the claim during the preview, details
and presentation screen. If the decoded claim is of type
`FiscalCodeClaim` then it extracts the fiscal code without the prefix.

## How to test
Test a eID issuance flow and check the preview screen. Then check the
details screen and try to obtain a MDL to check the presentation screen.
The fiscal code shouldn't have the `TINIT` prefix;