chore: [IOBP-787] No payment method screen during payment flow (#6005)
## Short description
This PR implements the logic if a user enters the payment method
selection screen but has no method onboarded. Before the "payment as
guest" will be implemented, a page should be shown inviting the user to
onboard a new card or proceed with other payment methods

## List of changes proposed in this pull request
- Added the request to get the user wallets when the user is in the
payment details screen page
- Redirect the user in a different screen if the user wallets list is
zero, otherwise (even if the API returns with an error) it should
navigate to the pick payment method screen. (Of course in case of error,
it will be shown an error screen)
- Edited the `notHasValidPaymentMethodsSelector` checking only the user
wallets (It will be removed as soon as the payment as guest feature is
implemented).

## How to test
Start a new payment flow with any method onboarded, you should be
redirected in a different screen that drives the user to onboard a new
card or choosing other payment method.

## Preview


https://github.com/user-attachments/assets/6d1ea49f-a8b2-4a86-b1d6-d8c3cec4d57f