chore(IT Wallet): [SIW-1384] Remove unique_id claim from eID (#6031)
> [!WARNING]
> Depends on #6030 

## Short description
This PR implements a dirty hack to temporary remove the `unique_id`
claim from the eID. A task to remove it has been already created and
mentioned in the code comments.

## List of changes proposed in this pull request
- Add a `removeClaim` utils which removes a claim from a
`ParsedCredential`. It returns the very same record if the claim doesn't
exists;
- Update the preview/details/presentation claim rendering to use
`removeClaim` on the `ParsedCredential`.

## How to test
Test the eID credential preview/details and presentation screen when
obtaining another credential. The `unique_id` claim should be hidden.

---------

Co-authored-by: Mario Perrotta <mario.perrotta@pagopa.it>