chore(IT Wallet): [SIW-1389] Generic error screen during wallet instance creation (#6035)
## Short description
This PR replaces the unsupported device error with a generic error when
a wallet instance is created. Currently the wallet instance creation
error does not provide enough information to ascertain the device is not
supported.

## List of changes proposed in this pull request
- Renamed the current generic error to issuer generic, because it hints
at problems on the issuer side
- Created a new generic error

## How to test
Fail the first step, where the wallet instance is created. You should
see the following screen:


![wallet_creation_generic_error](https://github.com/user-attachments/assets/33809394-640e-4cf3-9c10-f0322b307c31)

---------

Co-authored-by: Federico Mastrini <federicomastrini93@gmail.com>
Co-authored-by: Mario Perrotta <mario.perrotta@pagopa.it>
Co-authored-by: LazyAfternoons <LazyAfternoons@users.noreply.github.com>