fix(Cross): [IOAPPX-363] Discard activity state persistence in Android (#6075)
## Short description
> On Android the View state is not persisted consistently across
Activity restarts, which can lead to crashes in those cases. It is
recommended to override the native Android method called on Activity
restarts in your main Activity, to avoid these crashes.

[Source](https://github.com/software-mansion/react-native-screens?tab=readme-ov-file#android)

## List of changes proposed in this pull request
- Replace `savedInstanceState` with `null` in `MainActivity.java` to
discard any persisted Activity state during the restart process. This
should resolve the crashes we are observing in production.
 
## How to test
```
> cd android && ./gradlew clean && cd .. 
> yarn react-native run-android
```
Test that the build works and that there's no regression in the app when
restarting it or when swapping between recent apps.

Co-authored-by: Andrea <andrea.piai@pagopa.it>