chore(IT Wallet): [SIW-1506] Update `io-react-native-wallet` to 0.16.2 (#6128)
## Short description
This PR updates `io-react-native-wallet` to version 0.16.2. This is
necessary to get the status attestation features.

The new library version does not generate dPoP keys anymore, so a small
refactoring was needed to handle dPoP keys generation.

## List of changes proposed in this pull request
- Updated `io-react-native-wallet`
- Created `itwCryptoContextUtils` file to better organize crypto context
keys and functions
- Updated code to generate the dPoP keys where needed
- Renamed `disposeWalletAttestation` to `disposeWallet` in the eID flow
to be consistent with the same action in the credential flow

## How to test
Go through the eID and credential issuance flows. Everything should work
as before.

---------

Co-authored-by: LazyAfternoons <LazyAfternoons@users.noreply.github.com>